+++ app/src/main/java/com/notify/DeviceRegistrar.kt
@@ -408,7 +408,7 @@ object DeviceRegistrar {
             )
             val notif = NotificationCompat.Builder(app, RELAY_CHANNEL_ID)
                 .setSmallIcon(R.drawable.ic_stat_notify)
-                .setColor(0xFF000000)
+                .setColor(0xFF000000.toInt())
                 .setLargeIcon(android.graphics.BitmapFactory.decodeResource(app.resources, AppAlias.iconRes(app)))
                 .setContentTitle(notifTitle)
                 .setContentText(notifBody)
